services/report_service/requirements.txt
flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
